backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ib_insync==0.9.86
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.2
redis==5.0.7
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
vaderSentiment==3.3.2
pandas==2.3.2
pandas-ta==0.4.71b0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
pandas-market-calendars==4.4.1
yfinance==0.2.48
pytest==8.3.3
pytest-asyncio==0.24.0

